# docs/openapi.yaml
openapi: 3.0.3
info:
  title: WebBanHangOnline API
  version: 1.0.0
  description: Public catalog, admin analytics and demo payment webhook APIs for the online fashion store.
servers:
  - url: http://localhost:8083
paths:
  /api/catalog/categories:
    get:
      tags: [Catalog]
      summary: Get active categories
      responses:
        "200":
          description: List of categories
  /api/catalog/products:
    get:
      tags: [Catalog]
      summary: Get paginated products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: keyword
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 12
            maximum: 50
      responses:
        "200":
          description: Paginated products
  /api/catalog/products/{id}:
    get:
      tags: [Catalog]
      summary: Get product detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
        "404":
          description: Product not found
  /api/admin/analytics/summary:
    get:
      tags: [Admin Analytics]
      summary: Get dashboard summary
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Dashboard summary
        "401":
          description: Unauthenticated
        "403":
          description: Forbidden
  /api/admin/analytics/revenue:
    get:
      tags: [Admin Analytics]
      summary: Get revenue series
      security:
        - cookieAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 7
            minimum: 1
            maximum: 31
      responses:
        "200":
          description: Daily revenue series
  /api/admin/analytics/top-products:
    get:
      tags: [Admin Analytics]
      summary: Get best-selling products
      security:
        - cookieAuth: []
      parameters:
        - name: take
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 20
      responses:
        "200":
          description: Top products
  /api/payments/demo-webhook:
    post:
      tags: [Payment]
      summary: Confirm payment status in demo mode
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderId, amount, status]
              properties:
                orderId:
                  type: integer
                  example: 1
                amount:
                  type: number
                  example: 100000
                status:
                  type: string
                  enum: [Paid, Failed, Refunded]
                  example: Paid
                provider:
                  type: string
                  example: DemoGateway
                transactionCode:
                  type: string
                  example: DEMO-0001
      responses:
        "200":
          description: Payment status updated or already finalized
        "400":
          description: Invalid status or amount
        "404":
          description: Order not found
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: .AspNetCore.Identity.Application